func-package/requirements.txt
azure-functions
requests
